Replaced example with actual VS projects
Removed the example file that was created from what was posted on the forum and replaced it with the actual Visual Studio Project.
Also added an updated version of the project that uses the IVI standard VISA.NET instead of NI-VISA.
