+++ Oscilloscopes/PerformanceScopes/src/CurvestreamExample/CurvestreamExample-IVI-VISA.NET/CurvestreamExample-IVI-VISA.NET.sln
@@ -1,0 +1,20 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 11.00
+# Visual Studio 2010
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CurvestreamExample-IVI-VISA.NET", "CurvestreamExample-IVI-VISA.NET\CurvestreamExample-IVI-VISA.NET.csproj", "{09335BF5-9F20-4FD5-9948-4F88ABBE45BB}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x86 = Debug|x86
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{09335BF5-9F20-4FD5-9948-4F88ABBE45BB}.Debug|x86.ActiveCfg = Debug|x86
+		{09335BF5-9F20-4FD5-9948-4F88ABBE45BB}.Debug|x86.Build.0 = Debug|x86
+		{09335BF5-9F20-4FD5-9948-4F88ABBE45BB}.Release|x86.ActiveCfg = Release|x86
+		{09335BF5-9F20-4FD5-9948-4F88ABBE45BB}.Release|x86.Build.0 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
