+++ Oscilloscopes/PerformanceScopes/src/CurvestreamExample/CurvestreamExample-IVI-VISA.NET/CurvestreamExample-IVI-VISA.NET/Properties/Settings.settings
@@ -1,0 +1,7 @@
+﻿<?xml version='1.0' encoding='utf-8'?>
+<SettingsFile xmlns="http://schemas.microsoft.com/VisualStudio/2004/01/settings" CurrentProfile="(Default)">
+  <Profiles>
+    <Profile Name="(Default)" />
+  </Profiles>
+  <Settings />
+</SettingsFile>
