+++ Oscilloscopes/PerformanceScopes/src/CurvestreamExample/CurvestreamExample-NI-VISA/CurvestreamExample/app.config
@@ -1,0 +1,11 @@
+<?xml version="1.0"?>
+<configuration>
+<startup><supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.0"/></startup>	<runtime>
+		<assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
+			<dependentAssembly>
+				<assemblyIdentity name="NationalInstruments.Common" publicKeyToken="DC6AD606294FC298" culture="neutral"/>
+				<bindingRedirect oldVersion="0.0.0.0-15.0.40.49154" newVersion="15.0.40.49154"/>
+			</dependentAssembly>
+		</assemblyBinding>
+	</runtime>
+</configuration>
